Reintroduce q_qwerty.txt mapping
With a laptop it's annoying to switch between Qwerty and Dvorak modes
when switching between the Kinesis and the laptop keyboard. To keep the
laptop in Dvorak mode, set the Kinesis in Qwerty mode and make a few
adjustments so the correct keys are sent (The [intl-\] and open- and
close brackets get a kind of double mapping which needs to be corrected.